+++ corelib/src/libs/SireMM/morsepotentialrestraints.cpp
@@ -362,11 +362,11 @@ static const RegisterMetaType<MorsePotentialRestraints> r_morsepotentialrests;
 
 QDataStream &operator<<(QDataStream &ds, const MorsePotentialRestraints &morser_morsepotentialrests)
 {
-    writeHeader(ds, r_morsepotentialrests, 1);
+    writeHeader(ds, r_morsepotentialrests, 2);
 
     SharedDataStream sds(ds);
 
-    sds << morser_morsepotentialrests.r
+    sds << morser_morsepotentialrests.r << morser_morsepotentialrests.use_pbc
         << static_cast<const Restraints &>(morser_morsepotentialrests);
 
     return ds;
@@ -382,8 +382,15 @@ QDataStream &operator>>(QDataStream &ds, MorsePotentialRestraints &morser_morsep
 
         sds >> morser_morsepotentialrests.r >> static_cast<Restraints &>(morser_morsepotentialrests);
     }
+    else if (v == 2)
+    {
+        SharedDataStream sds(ds);
+
+        sds >> morser_morsepotentialrests.r >> morser_morsepotentialrests.use_pbc
+            >> static_cast<Restraints &>(morser_morsepotentialrests);
+    }
     else
-        throw version_error(v, "1", r_morsepotentialrests, CODELOC);
+        throw version_error(v, "1,2", r_morsepotentialrests, CODELOC);
 
     return ds;
 }
@@ -436,7 +443,7 @@ MorsePotentialRestraints::MorsePotentialRestraints(const QString &name,
 }
 
 MorsePotentialRestraints::MorsePotentialRestraints(const MorsePotentialRestraints &other)
-    : ConcreteProperty<MorsePotentialRestraints, Restraints>(other), r(other.r)
+    : ConcreteProperty<MorsePotentialRestraints, Restraints>(other), r(other.r), use_pbc(other.use_pbc)
 {
 }
 
@@ -447,13 +454,14 @@ MorsePotentialRestraints::~MorsePotentialRestraints()
 MorsePotentialRestraints &MorsePotentialRestraints::operator=(const MorsePotentialRestraints &other)
 {
     r = other.r;
+    use_pbc = other.use_pbc;
     Restraints::operator=(other);
     return *this;
 }
 
 bool MorsePotentialRestraints::operator==(const MorsePotentialRestraints &other) const
 {
-    return r == other.r and Restraints::operator==(other);
+    return r == other.r and Restraints::operator==(other) and use_pbc == other.use_pbc;
 }
 
 bool MorsePotentialRestraints::operator!=(const MorsePotentialRestraints &other) const
@@ -507,7 +515,11 @@ QString MorsePotentialRestraints::toString() const
         }
     }
 
-    return QObject::tr("MorsePotentialRestraints( name=%1, size=%2\n%3\n)").arg(this->name()).arg(n).arg(parts.join("\n"));
+    return QObject::tr("MorsePotentialRestraints( name=%1, size=%2, use_pbc=%3\n%4\n)")
+            .arg(this->name())
+            .arg(n)
+            .arg(this->use_pbc ? "use_pbc=True" : "use_pbc=False")
+            .arg(parts.join("\n"));
 }
 
 /** Return whether or not this is empty */
@@ -689,4 +701,16 @@ MorsePotentialRestraints MorsePotentialRestraints::operator+(const MorsePotentia
     MorsePotentialRestraints ret(*this);
     ret += restraints;
     return *this;
+}
+
+/** Set whether or not periodic boundary conditions are to be used */
+void MorsePotentialRestraints::setUsesPbc(bool use_pbc)
+{
+    this->use_pbc = use_pbc;
+}
+
+/** Return whether or not periodic boundary conditions are to be used */
+bool MorsePotentialRestraints::usesPbc() const
+{
+    return this->use_pbc;
 }